Refactor memory queue receiver code to use a memory queue client API for better readability.
