+++ src/lib.rs
@@ -3,11 +3,10 @@
 mod config_event;
 mod topology_controller;
 mod model;
+mod memory_queue_client;
 
-use tracing::info;
-use vector::event::EventContainer;
-use vector::topology::GLOBAL_VEC_RX;
 use crate::topology_controller::TopologyController;
+use crate::memory_queue_client::MemoryQueueClient;
 use crate::model::CxxLogEvent;
 
 #[cxx::bridge(namespace = "vectorcxx")]
@@ -41,8 +40,6 @@ mod ffi {
          */
         type CxxLogEvent;
 
-        fn poll_vector_events() -> Vec<CxxLogEvent>;
-
         unsafe fn get<'a>(self: &'a CxxLogEvent, key: &str) -> &'a str;
 
         // return a field as timestamp with microsecond precision
@@ -50,40 +47,23 @@ mod ffi {
 
         fn fields(self: &CxxLogEvent) -> Vec<String>;
     }
-}
 
-pub fn poll_vector_events() -> Vec<CxxLogEvent> {
-    let out_rx = unsafe { &mut GLOBAL_VEC_RX };
-    let mut events_list: Vec<CxxLogEvent> = Vec::new();
+    extern "Rust" {
+        /**
+         * MemoryQueueClient
+         */
+        type MemoryQueueClient;
 
-    if let Some(rx) = out_rx {
-        match rx.try_next() {
-            Ok(Some(value)) => {
-                // info!("\nevent is {:?}", value[0].as_log());
-                // if let Some(target) = value[0].as_log().get("-Target-Es") {
-                //     target_es = target.to_string_lossy();
-                // } else if let Some(target) = value[0].as_log().get("_target_es") {
-                //     target_es = target.to_string_lossy();
-                // }
-                info!("event array received length={:?}", value.len());
-                events_list.reserve(value.len());
-                value.iter_events().for_each(|event_ref|
-                    events_list.push(CxxLogEvent { log_event: event_ref.into_log() })
-                );
-            }
-            Ok(None) => {
-                // std::thread::sleep(ten_millis);
-                // info!("polling channel closed");
-            }
-            Err(_) => {
-                // std::thread::sleep(ten_millis);
-                // info!("polling vector events error: {:?}\n", e);
-            }
-        }
+        fn new_memory_queue_client() -> Box<MemoryQueueClient>;
+
+        fn poll(self: &mut MemoryQueueClient) -> Vec<CxxLogEvent>;
     }
-    events_list
 }
 
 pub fn new_topology_controller() -> Box<TopologyController> {
     Box::new(TopologyController::new())
+}
+
+pub fn new_memory_queue_client() -> Box<MemoryQueueClient> {
+    Box::new(MemoryQueueClient::new())
 }
