--- src/main.cpp
@@ -1,7 +1,0 @@
-// we need a cpp file, or cmake will give 'No SOURCES given to target' error when it try to target sources.
-#include "vectorcxx/src/lib.rs.h"
-
-namespace vectorcxx {
-    void start() {
-    }
-}
