+++ src/vectorcxx/cxx_memory_queue_client.h
@@ -1,0 +1,10 @@
+
+#pragma once
+
+#include "vectorcxx/src/lib.rs.h"
+namespace vectorcxx {
+  class CxxMemoryQueueClient {
+  public:
+    static rust::Box<MemoryQueueClient> &get_instance();
+  };
+}
